requirements.txt
streamlit == 0.62.1
torch == 1.6.0
numpy == 1.18.1
Pillow == 7.0.0

